Ampere Hardraade:
Majority of the changes that I have made are bug fixes: removing run-time
errors, logic errors, and so forth.  Unfortunately, since the scripts are so
HUGE, I won't even dare to say that they are "bug free".

converter.nas has been added to the list of files.  What this Nasal script
does is to convert values from one base to another.  Other additions are the
'Legacy' folders; one in the 'Bus' directory and another in the 'Switch'
directory.  They are the old scripts before I switched medium from the
property tree to Nasal.  They may become handy in the future. =)