Ampere wrote a list:
See the attachment for a new Nasal script for the A380.  Please create a new
directory called "ADT" under the "Nasal" folder, and place this file there.
In addition, please remove the files "normal.nas" and "test.nas" from the
"Nasal" directory.

Description of the file:

This file is a doubly-linked-list data type implemented in Nasal.