Ampere K:
"All credits go to Fahim Dalvi for this impressive work.  This update adds
functional instruments and animations to the A380.  Currently, the flight
instruments are those from the 787, but in the future, they will be updated
to give a closer resemblance to Airbus cockpit.  Also, be advised that the
aircraft at this stage cannot be landed via VFR.  ILS approach must be used
due to line-of-sight issues."

/me: this commit leaves the A380 in a broken state, which isn't much worse
     than the previous state. :-)